fix(Arquitetura,UsuárioRepository): Ajustando arquitetura e repository
