+++ src/main/java/com/spring/ApiSystem/dto/usuario/request/EditarUsuarioDTO.java
@@ -1,20 +1,17 @@
-package com.spring.ApiSystem.interfaces.dto.usuario;
+package com.spring.ApiSystem.dto.usuario.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-public class UsuarioDTO {
-
-    @NotBlank(message = "O nome é obrigatório")
+public class EditarUsuarioDTO {
     private String nome;
 
-    @NotBlank(message = "O email é obrigatório")
-    @Email(message = "O email deve ser válido")
+    @Email(message = "Email deve ter formato válido")
     private String email;
 
-    @NotBlank(message = "A senha é obrigatória")
+    @NotBlank(message = "A senha deve ser válida")
     @Size(min = 6, message = "A senha deve ter no mínimo 6 caracteres")
     private String senha;
 
@@ -22,9 +19,9 @@ public class UsuarioDTO {
     @Size(min = 11, message = "O CPF deve conter 11 dígitos")
     private String cpf;
 
-    public UsuarioDTO() {}
+    public EditarUsuarioDTO() {}
 
-    public UsuarioDTO(String nome, String email, String senha, String cpf) {
+    public EditarUsuarioDTO(String nome, String email, String senha, String cpf) {
         this.nome = nome;
         this.email = email;
         this.senha = senha;
