--- src/main/java/com/spring/ApiSystem/infrastructure/jpa/user/JpaUsuarioRepository.java
@@ -1,14 +1,0 @@
-package com.spring.ApiSystem.infrastructure.jpa.user;
-
-import com.spring.ApiSystem.domain.entity.Usuario;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface JpaUsuarioRepository extends JpaRepository<Usuario,Long> {
-
-    Optional<Usuario> findByEmail(String email);
-    Optional<Usuario> findByCpf(String cpf);
-    Optional<Usuario> findById(Long id);
-
-}
